pool transactions component splitup
