+++ src/components/molecules/PoolTransactions/Title.module.css
@@ -1,0 +1,3 @@
+.titleText {
+  white-space: pre;
+}
