+++ src/components/molecules/PoolTransactions/index.module.css
@@ -1,7 +1,3 @@
 .time {
   color: var(--color-secondary);
 }
-
-.titleText {
-  white-space: pre;
-}
